remove uncessary eslint-disables, try to fix some catch statements, refactor food weight table into own component
