+++ zoo_frontend/src/pages/food/dataSrc/dataSrc.jsx
@@ -42,7 +42,7 @@ class Home extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      token: props.token, // eslint-disable-line react/no-unused-state
+      token: props.token,
       dataSrc: props.dataSrc,
     };
     this.notificationsRef = React.createRef();
